Keccak: reduce memory footprint
Reapply 070fb3410e82e21e32d2b2fb16722063ce170cb4 and modify it to support big-endian
systems.

This change allows the reenablement of the kernel crypto API support for SHA-3, SHAKE,
and KMAC.

+++ hash/src/keccak_asm_glue.h
@@ -397,7 +397,7 @@ static inline void keccak_asm_squeeze(
 			ExtractBytes(ctx->state, digest, ctx->offset,
 				     partialBlock);
 			digest += partialBlock;
-			ctx->offset += partialBlock;
+			ctx->offset += (uint8_t)partialBlock;
 		}
 	}
 }
